+++ savage/src/main.rs
@@ -6,11 +6,75 @@ mod input;
 use std::{collections::HashMap, fs};
 
 use ansi_term::Style;
+use ariadne::{Color, Fmt, Label, Report, ReportKind, Source};
 use directories::ProjectDirs;
 use rustyline::{error::ReadlineError, highlight::Highlighter, Editor};
-use savage_core::expression::{Expression, Vector};
+use savage_core::{
+    expression::{Expression, Vector},
+    parse::{Error, ErrorReason},
+};
 
 use crate::input::InputHelper;
+
+fn format_parse_error(error: Error<char>) -> Report {
+    // Heavily based on https://github.com/zesterer/chumsky/blob/463226372cf293d45bd5df52bf25d5028243066e/examples/json.rs#L114-L173
+    let message = if let ErrorReason::Custom(message) = error.reason() {
+        message.clone()
+    } else {
+        format!(
+            "{}, expected {}",
+            if error.found().is_some() {
+                "Unexpected token"
+            } else {
+                "Unexpected end of input"
+            },
+            if error.expected().len() == 0 {
+                "something else".to_string()
+            } else {
+                error
+                    .expected()
+                    .map(|expected| match expected {
+                        Some(expected) => expected.to_string(),
+                        None => "end of input".to_string(),
+                    })
+                    .collect::<Vec<_>>()
+                    .join(", ")
+            },
+        )
+    };
+
+    let report = Report::build(ReportKind::Error, (), error.span().start)
+        .with_message(message)
+        .with_label(
+            Label::new(error.span())
+                .with_message(match error.reason() {
+                    ErrorReason::Custom(message) => message.clone(),
+                    _ => format!(
+                        "Unexpected {}",
+                        error
+                            .found()
+                            .map(|c| format!("token {}", c.fg(Color::Red)))
+                            .unwrap_or_else(|| "end of input".to_string()),
+                    ),
+                })
+                .with_color(Color::Red),
+        );
+
+    let report = match error.reason() {
+        ErrorReason::Unclosed { span, delimiter } => report.with_label(
+            Label::new(span.clone())
+                .with_message(format!(
+                    "Unclosed delimiter {}",
+                    delimiter.fg(Color::Yellow),
+                ))
+                .with_color(Color::Yellow),
+        ),
+        ErrorReason::Unexpected => report,
+        ErrorReason::Custom(_) => report,
+    };
+
+    report.finish()
+}
 
 fn main() {
     let history_path = ProjectDirs::from("com.worldwidemann", "", "Savage")
@@ -81,7 +145,13 @@ fn main() {
                         }
                         Err(error) => println!("Error: {:#?}", error),
                     },
-                    Err(error) => println!("Error: {:#?}", error),
+                    Err(errors) => {
+                        for error in errors {
+                            format_parse_error(error)
+                                .print(Source::from(line))
+                                .expect("unable to print parse error");
+                        }
+                    }
                 }
             }
             Err(ReadlineError::Interrupted | ReadlineError::Eof) => {