+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+#### REPL
+
+- Proper formatting for parse errors
+
 ### Changed
 
 ### Fixed